refactor(notifier): encapsulates sc interaction for edit/create into a single component
